Add functionParameters parameter to transform function
This parameter will make a set of user defined parameters available to
the transform function.

The purpose is to allow the user to specify a set of data to be used in
our preconfigured functions. For example, the mapping of codes to values
will require the user to provide a mapping of codes to values which we
will then use to map fields inplementing this function.

This makes the openhim-mediator-mapping service more configurable by the
user.

OHM-878