Add specifying logic to one to many transform
The previous function relied on the order of the input items. This made
the function sensitive to any changes in the mapping schema :( Not
ideal.

Now the function is far more robust. It applies regex functions to
ensure good data quality and to find the appropriate array to
manipulate.

OHIE-301