Add object pass through function to transforms
When mapping data the mapping syntax always moves data from a highlevel
object to a nested position within the output. This is fine for
individual fields. However, if you want the entire input object pass
through as the high level output object this would be a manual process
of moving each field over to copy the input object structure.
This transform function solves this by assigning the input object to the
output object thereby keeping it on the same object level hierachy

DEB-20